Reduce code smells in service objects

+++ service/clear_old.rb
@@ -1,14 +1,6 @@
 class ClearOld
-  def initialize(done_arr)
-    @done = done_arr
-  end
-
-  def clear
-    @done.each do |task|
-      diff = (Date.today - Date.parse(task.due)).to_i
-      if diff > 30
-        task.destroy
-      end
-    end
+  def self.clear(task)
+    diff = (Date.today - Date.parse(task.due)).to_i
+    task.destroy if diff > 30
   end
 end
